+++ test/test-main.hs
@@ -7,6 +7,7 @@ import TestParallelReduction
 import PropOpenClose
 import TinyLam
 import TestACompare
+import TestShiftEmbed
 import TestTH
 
 main :: IO ()
@@ -17,5 +18,6 @@ main = defaultMain $ testGroup "unboundGenerics"
        , test_openClose
        , test_tinyLam
        , test_acompare
+       , test_shiftEmbed
        , test_TH
        ]